Moved network out of controller class Changed direction class a bit
