+++ src/ESnake/controllers/__init__.py
@@ -1,1 +1,5 @@
+from .activations import Activations
+from .netv2 import NeuronV2, LayerV2
+from .netv2 import NetV2
+from .automatic import Automatic
 from .dylan import Dylan
